+++ CHANGELOG.md
@@ -79,6 +79,14 @@ This is the rolling changelog for TShock for Terraria. Use past tense when addin
 * Removed `Utils.GetGroup()` also have you seen `Groups.GetGroupByName()`? (@hakusaro)
 * `Utils.MaxChests()` is now `Utils.HasWorldReachedMaxChests()`. (@hakusaro)
 * `Utils.GetIPv4Address()` is now `Utils.GetIPv4AddressFromHostname()`. (@hakusaro)
+* Fixed the disappearing problem when placing tile entities. (@mistzzt)
+* Removed the stat tracking system. (@hakusaro)
+* Fixed erroneous kicks and bans when using `KickOnMediumcoreDeath` and `BanOnMediumcoreDeath` options. (@DankRank)
+* Removed `TSPlayer.InitSpawn` field. (@DankRank)
+* `OnPlayerSpawn`'s player ID field is now `PlayerId`. (@DankRank)
+* `Utils.TryParseTime` can now take spaces (e.g., `3d 5h 2m 3s`) (@QuiCM)
+* Enabled banning unregistered users (@QuiCM)
+* Added filtering and validation on packet 96 (Teleport player through portal) (@QuiCM)
 
 ## TShock 4.3.25
 * Fixed a critical exploit in the Terraria protocol that could cause massive unpreventable world corruption as well as a number of other problems. Thanks to @bartico6 for reporting. Fixed by the efforts of @QuiCM, @hakusaro, and tips in the right directioon from @bartico6.